ElectroPlus-Gateway/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.28.0
gunicorn==23.0.0
python-dotenv==1.1.1
requests==2.32.5
responses==0.25.2
pytest==8.3.2